Refactor api to two new models
the api.js file should only hold the api routes and not any of the logics. fileService handles files, folderService handles folders. Simple :-)
